BACKPORT: sched/fair: if sync flag ignored, try to place in mid cluster
If the sync flag is ignored because the current cpu is not in
the affinity mask for the target of a sync wakeup (usually
binder call), prefer to place in the mid cluster if
possible. The main case is a "top-app" task waking a
"foreground" task when the top-app task is running on
a CPU that is not in the foreground cpuset. This patch
causes the search order from mid capacity cpu be used
when the sync flag failed.

backport from commit 98ae57d9eaf7
("ANDROID: sched/fair: if sync flag ignored, try to place in same cluster")

Bug: 117438867
Test: boot to home, operation normal
Change-Id: I68d0cc05db1bc2cb02d4445c71b02215209e8c04
Signed-off-by: Rick Yiu <rickyiu@google.com>
Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>
Move definition of sync_boost up to accommodate for missing [1].

[1]: https://github.com/aosp-mirror/kernel_msm/commit/cea78e291fda0
Above revert was not merged as it caused misbehavior of prime core
where it would often jump to higher frequencies even at 0% load.

Signed-off-by: alk3pInjection <webmaster@raspii.tech>
(cherry picked from commit 29cd70f4c570f390aac6d18336f824f32f64c453)
Signed-off-by: Zlatan Radovanovic <zlatan.radovanovic@fet.ba>